requirements.txt
cffi==1.14.6
cryptography==3.4.7
greenlet==1.1.0
importlib-metadata==4.6.1
numpy==1.21.1
pandas==1.3.1
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
tabulate==0.8.9
typing-extensions==3.10.0.0
zipp==3.5.0
